small: fix assertion during slab allocation
When `slab_get` is called from `region_alloc` or `ibuf_alloc` first
of all we try to calculate order of appropriate slab. But there is
no check that requested size with slab meta is <= UINT32_MAX, that
leads to assertion failer in `slab_order` function. There is no
need for this assertion we should return `cache->order_max + 1`
for this case.

Closes #6726