evio: pass iostream instead of fd (and uri) to on_accept callback
Currently, we always create a plain stream, but soon there will be an
option to create an encrypted stream so let's pass iostream, which will
encapsulate this information, instead of fd to the accept callback. The
iostream will be created by evio in accordance with the settings passed
in uri parameters so uri won't be needed in the accept callback.