msgpack: fixed incorrect print of msgpack field
Due to the fact that we used `const char**` to save a pointer to the
original msgpack header, when restoring the pointer, we received an
incorrect value and as a consequence assertion failure later. Correctly
saving and restoring the pointer fixed this problem.

NO_DOC=bug fix
NO_CHANGELOG=bug fix