iostream: introduce context and use it for creating streams
IO stream context (iostream_ctx) stores information needed to construct
an iostream object.  A context is created from URI and mode, which is
either server or client. Currently, both parameters are ignored and a
newly created context always constructs plain streams. However, in
future we will create SSL streams for certain URIs. That's when we'll
need the mode (SSL connection is asymmetrical).