memtx: add tree_iterator_set_current helper
So as not to set tree_iterator::current directly. Currently, we only
need to ref/unref tuple when we update current, but in the future we
will also alloc/free func index key part, a pointer to which is stored
in a hint.

Needed for #6786

(cherry picked from commit d8a352b934507c59d934782e1ae490e5fecfb2f1)