Rework `index_get` and `iterator_next`functions
Currently `index_get` and `iterator_next` functions return
result tuple as is, without any transformation, because
tuple is stored in the space in the same format in which
it is returned to the user. Later after tuple compression
implementation (and maybe some other features) tuple should
be transformed before return it to user. Rework `index_get`
and `iterator_next` functions - now they are expected to
return the tuple in format appropriate for user. Implement
`index_get_raw` and `iterator_next_raw` functions, which
return tuple as is.

Part of #2695

NO_CHANGELOG=internal changes
NO_DOC=internal changes