applier: remove useless state check from writer
Applier writer fiber is started after JOIN so the only possible states
are READY, SYNC, and error states. It's fine to try to write on error -
the operation will be silently ignored (due to EPIPE check below).

This is a step towards moving the applier fiber from the main to the
applier thread, which is required to properly handle SSL connections.

NO_DOC=refactoring
NO_CHANGELOG=refactoring