raft: send promote only after new term has been broadcasted
Store last sent term for each relay and wait for it to be
greater or equal to new term before writing promote
or demote to wal. Before it was possible that promote has been written
to wal, read and broadcasted by relay before term increment was
processed by relay.

Closes #6754
Part of #6754

NO_DOC=Bug fix
NO_TEST=In previous commit