console: introduce __autocomplete metamethod
This patch updates the logic in lua_rl_complete() where __index from
a metatable apparently expected to be a table, not a function. It is
rather simple for a table object, but some of them are userdata, so
an __autocomplete method is introduced to be sure it returns a table
with all names suitable for tabcompletion.

Part of #6305
NO_DOC=internal
NO_CHANGELOG=part of a fix