ci: not run workflows in forks
We have recently moved all development from local branches to forks.
Most workflows cannot be run in forks because many of them need access
to private runners and/or secrets.

Workflows in forks can be disabled manually, but it's hard to remember
and it needs an extra action from a developer. Instead, we can detect
the repo in workflows and run them if the repo is tarantool/tarantool.

Closes #6913

NO_DOC=ci
NO_CHANGELOG=ci
NO_TEST=ci