txm: fix a crash in mvcc with secondary index conflict
Memtx TX manager stores a pointer to deleting statement in prepared
story. This pointer is set in two cases:
1. a statement deletes (or overwrites) a tuple
2. a story becomes prepared while other inprogress TX overwrites it

By design a tuple can be deleted only by primary index, the case
when a transaction overwrites somehting in secondary index but does
not overwrite the same tuple in primary index is prohibited. That's
why the pointer (to deleting statement) must be set by and only by
the next statement in the primary index chain.

By mistake the pointer is set also in second index chain analysis
after reordering which led to unexpected state of a story.

The patch removes the problem.

Closes #6452
NO_DOC=bugfix