ci: run freebsd only on PRs with 'full-ci' label
It looks like we just missed the freebsd-12.yml and freebsd-13.yml
workflows when cherry picking this functionality from the master branch.

Now to run FreeBSD tests, one should either name the branch `*-full-ci`
and push it to the main repository or set the 'full-ci' label on the
pull request.

Follows up #6605