sql: introduce built-in function NOW()
Part of #6773

NO_CHANGELOG=It will be added in another commit.

@TarantoolBot document
Title: SQL built-in function NOW()

The NOW() function returns the current date and time as a DATETIME
value. If this function is called more than once in a query, it returns
the same result until the query completes, unless a yield has occurred.
On yield the value returned by NOW() changes.