sql: add support of DECIMAL and INTEGER to ROUND()
This patch fixes an issue with the implicit cast of INTEGER and DECIMAL
values to DOUBLE when they are passed as the first argument to the
ROUND() function.

Closes #6988

@TarantoolBot document
Title: ROUND() now properly supports INTEGER and DECIMAL

INTEGER and DECIMAL values passed as the first argument now will not be
cast to DOUBLE and the result will be of the same type as the first
argument. Also, the default type for the ROUND() is now DECIMAL.