mvcc: track nothing-found range read
MVCC must track a read set of transaction, including cases when
the select returns an empty result. In this case MVCC has two
options: to store a point key if the select has a full key or
to store a gap otherwise.

By a mistake the second variant was not implemented. Fix it now.

Closes #7025

NO_DOC=bugfix