raft: introduce Pre-Vote machinery
Support Pre-Vote in Raft: when a node knows, that some of the peers see
the leader of the current term, it doesn't start elections. Even when it
doesn't have a connection to the leader.

Don't use the machinery yet: nodes never share info on whether they see
the leader.

Part-of #6654

NO_DOC=no user-visible changes
NO_CHANGELOG=no user-visible changes