sql: check left operand before right in arithmetic
Prior to this patch, the right operand was checked before the left. This
can lead to some complications when arithmetic operations are added for
the DATETIME and INTERVAL values. To avoid this, let's check the left
operand before the right one. This will cause the description of the
error to change, but nothing else.

Needed for #6773

NO_DOC=Not documented feature.