sql: arithmetic for DATETIME and INTERVAL
After this patch, it will be possible to perform permitted arithmetic
operations on DATETIME and INTERVAL values.

Closes #6773

@TarantoolBot document
Title: Field type interval is available in SQL

The INTERVAL field type is now available in SQL. INTERVAL values cannot
be compared. INTERVAL values can be explicitly and implicitly cast to
ANY. There are currently no literals for the INTERVAL field type.