sql: drop extern_alloc flag of sql_expr_delete
It's always false.

I don't understand why it was added in the first place. Initially, it
was introduced by commit b9c9792d1ab95 ("sql: move default col values to
Tarantool's core"). The flag was set to true only in Tarantool core for
some reason. Then there was commit 7c6c572c9f44d ("sql: Fix DEFAULTs AST
memory leak on alter") that claimed there was a memory leak because the
flag was set it to true and switched it to false in all places except
one: when a space_def was destroyed it was still set to true. I dropped
usage of space_expr_dup and switched the flag to false in the previous
commit, and everything seems to work just fine now.

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring