alter: raise ER_WRONG_SPACE_FORMAT instead of ER_CREATE/ALTER_SPACE
This is needed to factor out the function that decodes space format so
that it can be used for online space upgrade. It's not a big deal,
because we already use ER_WRONG_SPACE_OPTIONS if space options specified
on space creation/alter are wrong so using ER_WRONG_SPACE_FORMAT would
only be consistent.

NO_DOC=minor change in error message
NO_CHANGELOG=minor change in error message