index: add iterator_space helper
The new helper function returns the space the iterator was created for.
It looks up the space by id and caches it in iterator struct. We need it
for space upgrade: we will use it in box_iterator_next() to update
tuples returned by iterator_next().

NO_DOC=internal
NO_TEST=internal
NO_CHANGELOG=internal