luatest: use /tmp/t as default socket dir
This patch changes the length of default socket paths for luatest tests
(when the VARDIR environment variable is not defined). It should reduce
the chance for the issue with the max length of socket paths (107 chars)
to happen.

NO_DOC=testing stuff
NO_TEST=testing stuff
NO_CHANGELOG=testing stuff